+++ zscaler/zia/zia_service.py
@@ -48,6 +48,7 @@ from zscaler.zia.dlp_resources import DLPResourcesAPI
 from zscaler.zia.dlp_templates import DLPTemplatesAPI
 from zscaler.zia.dlp_web_rules import DLPWebRuleAPI
 from zscaler.zia.dns_gatways import DNSGatewayAPI
+from zscaler.zia.email_profiles import EmailProfilesAPI
 from zscaler.zia.end_user_notification import EndUserNotificationAPI
 from zscaler.zia.file_type_control_rule import FileTypeControlRuleAPI
 from zscaler.zia.forwarding_control import ForwardingControlAPI
@@ -92,6 +93,7 @@ from zscaler.zia.vzen_clusters import VZENClustersAPI
 from zscaler.zia.vzen_nodes import VZENNodesAPI
 from zscaler.zia.workload_groups import WorkloadGroupsAPI
 from zscaler.zia.zpa_gateway import ZPAGatewayAPI
+from zscaler.zia.email_profiles import EmailProfilesAPI
 
 
 class ZIAService:
@@ -743,3 +745,11 @@ class ZIAService:
 
         """
         return SecureBrowsingAPI(self._request_executor)
+    
+    @property
+    def email_profiles(self) -> "EmailProfilesAPI":
+        """
+        The interface object for the :ref:`ZIA Email Profiles API interface <zia-email_profiles>`.
+
+        """
+        return EmailProfilesAPI(self._request_executor)